// front-medico/src/generated/openapi/model/nuevaPrescripcion.ts
/**
 * Imani API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NuevaPrescripcion { 
    medicamento: string;
    frecuenciaDosis: number;
    indicaciones: string;
    duracion: number;
    inicio?: string;
    identificador?: string;
    figura: NuevaPrescripcion.FiguraEnum;
    /**
     * La cantidad que tiene la caja o la cantidad que le queda al paciente en estos momentos 
     */
    numeroDeDosis: number;
    /**
     * cuanto tiene que tomar cada vez 
     */
    cantidadPorDosis: number;
}
export namespace NuevaPrescripcion {
    export type FiguraEnum = 'cuadrado' | 'circulo' | 'triangulo' | 'estrella' | 'anillo' | 'semicirculo' | 'nube';
    export const FiguraEnum = {
        Cuadrado: 'cuadrado' as FiguraEnum,
        Circulo: 'circulo' as FiguraEnum,
        Triangulo: 'triangulo' as FiguraEnum,
        Estrella: 'estrella' as FiguraEnum,
        Anillo: 'anillo' as FiguraEnum,
        Semicirculo: 'semicirculo' as FiguraEnum,
        Nube: 'nube' as FiguraEnum
    };
}